HIP-136 Try getting app to load after db migration

+++ src/Traffk.Bal/Data/Rdb/ApplicationUser.cs
@@ -26,7 +26,7 @@ namespace Traffk.Bal.Data.Rdb
         public Tenant Tenant { get; set; }
 
         [Column("ContactId")]
-        public long ContactId { get; set; }
+        public int ContactId { get; set; }
 
         [ForeignKey("ContactId")]
         [JsonIgnore]
